Code review changes

+++ src/app/ssh-keys/redux/ssh-key.effects.ts
@@ -7,9 +7,7 @@ import { SSHKeyPair } from '../../shared/models/ssh-keypair.model';
 import { SshPrivateKeyDialogComponent } from '../ssh-key-creation/ssh-private-key-dialog.component';
 import { MatDialog } from '@angular/material';
 import { DialogService } from '../../dialog/dialog-service/dialog.service';
-import { sshKeyId } from './ssh-key.reducers';
 import { Router } from '@angular/router';
-import { Utils } from '../../shared/services/utils/utils.service';
 
 import * as sshKey from './ssh-key.actions';
 
@@ -30,31 +28,34 @@ export class SshKeyEffects {
   removeSshKeyPair$: Observable<Action> = this.actions$
     .ofType(sshKey.SSH_KEY_PAIR_REMOVE)
     .switchMap((action: sshKey.RemoveSshKeyPair) => {
-      return this.sshKeyService.remove({
-        name: action.payload.name,
-        account: action.payload.account,
-        domainid: action.payload.domainid
-      })
-        .concatMap(() => [
-          new sshKey.RemoveSshKeyPairSuccessAction(sshKeyId(action.payload)),
-          new sshKey.RemoveSshKeyPairSuccessNavigation(action.payload),
-        ])
-        .catch((error: Error) => {
-          return Observable.of(new sshKey.RemoveSshKeyPairErrorAction(error));
+      return this.dialogService.confirm({ message: 'SSH_KEYS.REMOVE_THIS_KEY' })
+        .onErrorResumeNext()
+        .filter(res => !!res)
+        .switchMap(() => {
+          return this.sshKeyService.remove({
+            name: action.payload.name,
+            account: action.payload.account,
+            domainid: action.payload.domainid
+          })
+            .map(() => new sshKey.RemoveSshKeyPairSuccessAction(action.payload))
+            .catch((error: Error) => {
+              return Observable.of(new sshKey.RemoveSshKeyPairErrorAction(error));
+            });
         });
     });
 
   @Effect({ dispatch: false })
-  removeSshKeyPairSuccessNavigate$: Observable<Action> = this.actions$
-    .ofType(sshKey.SSH_KEY_PAIR_REMOVE_SUCCESS_NAVIGATION)
-    .map((action: sshKey.RemoveSshKeyPairSuccessNavigation) => action.payload)
-    .do((res) => {
-      const route = Utils.deepestActivatedRoute(this.router);
-      if (route.snapshot.params.id === res.name && route.snapshot.queryParams.account === res.account) {
-        this.router.navigate(['./ssh-keys'], {
-          queryParamsHandling: 'preserve'
-        });
-      }
+  removeSshKeyPairSuccessNavigate$: Observable<SSHKeyPair> = this.actions$
+    .ofType(sshKey.SSH_KEY_PAIR_REMOVE_SUCCESS)
+    .map((action: sshKey.RemoveSshKeyPairSuccessAction) => action.payload)
+    .filter((sshKey: SSHKeyPair) => {
+      return this.router.isActive(`/ssh-keys/view/${sshKey.name}`, false)
+        && this.router.routerState.root.snapshot.queryParams.account === sshKey.account;
+    })
+    .do(() => {
+      this.router.navigate(['./ssh-keys'], {
+        queryParamsHandling: 'preserve'
+      });
     });
 
   @Effect()
